# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CareAI Healthcare Triage API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterInput {
  email: string;
  /** @minLength 6 */
  password: string;
  /** @minLength 1 */
  name: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface RefreshInput {
  refreshToken: string;
}

export interface User {
  id: number;
  email: string;
  name: string;
  createdAt: string;
}

export interface AuthResponse {
  accessToken: string;
  refreshToken: string;
  user: User;
}

export type UserProfileLanguage = typeof UserProfileLanguage[keyof typeof UserProfileLanguage];


export const UserProfileLanguage = {
  en: 'en',
  hi: 'hi',
  kn: 'kn',
} as const;

export interface UserProfile {
  id: number;
  email: string;
  name: string;
  language: UserProfileLanguage;
  voiceEnabled: boolean;
  voiceRate?: number;
  createdAt: string;
}

export type UserPreferencesInputLanguage = typeof UserPreferencesInputLanguage[keyof typeof UserPreferencesInputLanguage];


export const UserPreferencesInputLanguage = {
  en: 'en',
  hi: 'hi',
  kn: 'kn',
} as const;

export interface UserPreferencesInput {
  language?: UserPreferencesInputLanguage;
  voiceEnabled?: boolean;
  voiceRate?: number;
}

export type ChatMessageInputLanguage = typeof ChatMessageInputLanguage[keyof typeof ChatMessageInputLanguage];


export const ChatMessageInputLanguage = {
  en: 'en',
  hi: 'hi',
  kn: 'kn',
} as const;

export interface ChatMessageInput {
  message: string;
  language: ChatMessageInputLanguage;
  /** @nullable */
  sessionId?: string | null;
  /** @nullable */
  userId?: number | null;
}

export type ChatMessageRole = typeof ChatMessageRole[keyof typeof ChatMessageRole];


export const ChatMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface ChatMessage {
  id: number;
  role: ChatMessageRole;
  content: string;
  language: string;
  /** @nullable */
  sessionId?: string | null;
  createdAt: string;
}

export type ChatResponseRiskLevel = typeof ChatResponseRiskLevel[keyof typeof ChatResponseRiskLevel];


export const ChatResponseRiskLevel = {
  low: 'low',
  moderate: 'moderate',
  high: 'high',
  emergency: 'emergency',
} as const;

export interface ChatResponse {
  message: string;
  riskScore: number;
  riskLevel: ChatResponseRiskLevel;
  intent: string;
  sessionId: string;
  emergencyDetected?: boolean;
  recommendations?: string[];
}

export type AssessmentInputLanguage = typeof AssessmentInputLanguage[keyof typeof AssessmentInputLanguage];


export const AssessmentInputLanguage = {
  en: 'en',
  hi: 'hi',
  kn: 'kn',
} as const;

export interface AssessmentInput {
  symptoms: string[];
  age: number;
  gender: string;
  painLevel: number;
  /** @nullable */
  additionalNotes?: string | null;
  language: AssessmentInputLanguage;
  /** @nullable */
  userId?: number | null;
  /** @nullable */
  sessionId?: string | null;
  /** @nullable */
  patientName?: string | null;
}

export type AssessmentRiskLevel = typeof AssessmentRiskLevel[keyof typeof AssessmentRiskLevel];


export const AssessmentRiskLevel = {
  low: 'low',
  moderate: 'moderate',
  high: 'high',
  emergency: 'emergency',
} as const;

export interface Assessment {
  id: number;
  symptoms: string[];
  age: number;
  gender: string;
  painLevel: number;
  riskScore: number;
  riskLevel: AssessmentRiskLevel;
  recommendations: string[];
  /** @nullable */
  additionalNotes?: string | null;
  /** @nullable */
  patientName?: string | null;
  language?: string;
  createdAt: string;
}

export interface ReportInput {
  assessmentId: number;
  /** @nullable */
  userId?: number | null;
  /** @nullable */
  sessionId?: string | null;
}

export interface Report {
  id: number;
  assessmentId: number;
  summary: string;
  diagnosis?: string;
  recommendations?: string[];
  assessment?: Assessment;
  createdAt: string;
}

export type VoiceTranscriptInputLanguage = typeof VoiceTranscriptInputLanguage[keyof typeof VoiceTranscriptInputLanguage];


export const VoiceTranscriptInputLanguage = {
  en: 'en',
  hi: 'hi',
  kn: 'kn',
} as const;

export interface VoiceTranscriptInput {
  transcript: string;
  language: VoiceTranscriptInputLanguage;
  /** @nullable */
  userId?: number | null;
  /** @nullable */
  sessionId?: string | null;
}

export interface VoiceTranscript {
  id: number;
  transcript: string;
  language: string;
  aiResponse: string;
  intent?: string;
  createdAt: string;
}

export interface Doctor {
  id: number;
  name: string;
  specialty: string;
  experience: number;
  rating: number;
  reviewCount: number;
  available: boolean;
  avatarEmoji?: string;
  /** @nullable */
  nextSlot?: string | null;
}

export type AppointmentInputType = typeof AppointmentInputType[keyof typeof AppointmentInputType];


export const AppointmentInputType = {
  video: 'video',
  chat: 'chat',
  callback: 'callback',
} as const;

export interface AppointmentInput {
  doctorId: number;
  type: AppointmentInputType;
  /** @nullable */
  userId?: number | null;
  /** @nullable */
  sessionId?: string | null;
  /** @nullable */
  preferredTime?: string | null;
}

export interface Appointment {
  id: number;
  doctorId: number;
  type: string;
  status: string;
  doctor?: Doctor;
  /** @nullable */
  scheduledAt?: string | null;
  createdAt: string;
}

export type AnalyticsSummaryRiskDistribution = {
  low?: number;
  moderate?: number;
  high?: number;
  emergency?: number;
};

export type AnalyticsSummarySymptomTrendItem = {
  date?: string;
  count?: number;
};

export interface AnalyticsSummary {
  totalAssessments: number;
  avgRiskScore: number;
  riskDistribution: AnalyticsSummaryRiskDistribution;
  recentAssessments: Assessment[];
  symptomTrend?: AnalyticsSummarySymptomTrendItem[];
}

export type GetChatHistoryParams = {
sessionId?: string;
limit?: number;
};

export type GetAssessmentHistoryParams = {
limit?: number;
};

export type ListReportsParams = {
limit?: number;
};

export type ListDoctorsParams = {
specialty?: string;
available?: boolean;
};