# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CareAI Healthcare Triage API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshInput"
      responses:
        "200":
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user info
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /chat/message:
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a message to AI chatbot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
  /chat/history:
    get:
      operationId: getChatHistory
      tags: [chat]
      summary: Get chat history for current session or user
      parameters:
        - name: sessionId
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Chat history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
  /assessment/submit:
    post:
      operationId: submitAssessment
      tags: [assessment]
      summary: Submit a symptom assessment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssessmentInput"
      responses:
        "201":
          description: Assessment result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assessment"
  /assessment/history:
    get:
      operationId: getAssessmentHistory
      tags: [assessment]
      summary: Get assessment history
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Assessment history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assessment"
  /assessment/{id}:
    get:
      operationId: getAssessment
      tags: [assessment]
      summary: Get a specific assessment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Assessment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assessment"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /report/generate:
    post:
      operationId: generateReport
      tags: [report]
      summary: Generate a health report from an assessment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "201":
          description: Generated report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
  /report/list:
    get:
      operationId: listReports
      tags: [report]
      summary: List reports for current user or session
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Reports list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    RegisterInput:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        name:
          type: string
          minLength: 1
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    RefreshInput:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    User:
      type: object
      required: [id, email, name, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        createdAt:
          type: string
          format: date-time
    AuthResponse:
      type: object
      required: [accessToken, refreshToken, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: "#/components/schemas/User"
...
    ChatMessageInput:
      type: object
      required: [message, language]
      properties:
        message:
          type: string
        language:
          type: string
          enum: [en, hi, kn]
        sessionId:
          type: ["string", "null"]
        userId:
          type: ["integer", "null"]
    ChatMessage:
      type: object
      required: [id, role, content, language, createdAt]
      properties:
        id:
          type: integer
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        language:
          type: string
        sessionId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ChatResponse:
      type: object
      required: [message, riskScore, riskLevel, intent, sessionId]
      properties:
        message:
          type: string
        riskScore:
          type: integer
        riskLevel:
          type: string
          enum: [low, moderate, high, emergency]
        intent:
          type: string
        sessionId:
          type: string
        emergencyDetected:
          type: boolean
        recommendations:
          type: array
          items:
            type: string
    AssessmentInput:
      type: object
      required: [symptoms, age, gender, painLevel, language]
      properties:
        symptoms:
          type: array
          items:
            type: string
        age:
          type: integer
        gender:
          type: string
        painLevel:
          type: integer
        additionalNotes:
          type: ["string", "null"]
        language:
          type: string
          enum: [en, hi, kn]
        userId:
          type: ["integer", "null"]
        sessionId:
          type: ["string", "null"]
        patientName:
          type: ["string", "null"]
    Assessment:
      type: object
      required: [id, symptoms, age, gender, painLevel, riskScore, riskLevel, recommendations, createdAt]
      properties:
        id:
          type: integer
        symptoms:
          type: array
          items:
            type: string
        age:
          type: integer
        gender:
          type: string
        painLevel:
          type: integer
        riskScore:
          type: integer
        riskLevel:
          type: string
          enum: [low, moderate, high, emergency]
        recommendations:
          type: array
          items:
            type: string
        additionalNotes:
          type: ["string", "null"]
        patientName:
          type: ["string", "null"]
        language:
          type: string
        createdAt:
          type: string
          format: date-time
    ReportInput:
      type: object
      required: [assessmentId]
      properties:
        assessmentId:
          type: integer
        userId:
          type: ["integer", "null"]
        sessionId:
          type: ["string", "null"]
    Report:
      type: object
      required: [id, assessmentId, summary, createdAt]
      properties:
        id:
          type: integer
        assessmentId:
          type: integer
        summary:
          type: string
        diagnosis:
          type: string
        recommendations:
          type: array
          items:
            type: string
        assessment:
          $ref: "#/components/schemas/Assessment"
        createdAt:
          type: string
          format: date-time
...