# lib/api-zod/src/generated/types/assessmentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CareAI Healthcare Triage API
 * OpenAPI spec version: 0.1.0
 */
import type { AssessmentInputLanguage } from './assessmentInputLanguage';

export interface AssessmentInput {
  symptoms: string[];
  age: number;
  gender: string;
  painLevel: number;
  /** @nullable */
  additionalNotes?: string | null;
  language: AssessmentInputLanguage;
  /** @nullable */
  userId?: number | null;
  /** @nullable */
  sessionId?: string | null;
  /** @nullable */
  patientName?: string | null;
}